fix: correct 4 failing vitest property tests
- transactions: use Math.fround() for fc.float min/max (32-bit constraint)
- components P8: deduplicate by field before asserting (last-write-wins)
- components P26: verify display is always fixed template, not substring check
- components P27: filter generator to only produce elements with accessible names

+++ frontend/src/test/transactions.test.ts
@@ -12,7 +12,7 @@ describe('Property 15 — Transaction list renders required fields', () => {
           fc.record({
             id: fc.uuid(),
             orderId: fc.uuid(),
-            amount: fc.float({ min: 0.01, max: 1_000_000 }),
+            amount: fc.float({ min: Math.fround(0.01), max: Math.fround(1_000_000) }),
             status: fc.constantFrom('Pending_Payment', 'Settled', 'Cancelled'),
             createdAt: fc.string(),
             expiresAt: fc.string(),
